adding data plane for v1 v2
